+++ examples/ingest/box/ingest.sh
@@ -1,6 +1,6 @@
 #!/usr/bin/env bash
 
-# Processes several files in a nested folder structure from box://utic-test-ingest-fixtures/ 
+# Processes several files in a nested folder structure from box://utic-test-ingest-fixtures/
 # through Unstructured's library in 2 processes.
 
 # Structured outputs are stored in box-output/
@@ -26,7 +26,7 @@ PYTHONPATH=. ./unstructured/ingest/main.py \
    box \
    --box-app-config "$BOX_APP_CONFIG_PATH" \
    --remote-url box://utic-test-ingest-fixtures \
-   --structured-output-dir box-output \
+   --output-dir box-output \
    --num-processes 2 \
    --recursive \
-   --verbose 
+   --verbose