+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+## 0.10.14
+
+### Enhancements
+
+* Update all connectors to use new downstream architecture
+  * New click type added to parse comma-delimited string inputs
+  * Some CLI options renamed
+
+### Features
+
+### Fixes
+
 ## 0.10.13
 
 ### Enhancements
@@ -8,6 +20,7 @@
 * Updated Hi-Res Metadata: PDFs and Images using Hi-Res strategy now have layout model class probabilities added ot metadata.
 * Updated the `_detect_filetype_from_octet_stream()` function to use libmagic to infer the content type of file when it is not a zip file.
 * Tesseract minor version bump to 5.3.2
+
 ### Features
 
 * Add Jira Connector to be able to pull issues from a Jira organization